Make the nonogram example produce unique solutions
We have to be careful when exposing arrays containing
nil elements because Sentient could assign different
values to those elements and consider them unique
solutions, even though those elements are not exposed
to the user.

Going forwards, we should make it so that Sentient
disregards nil elements when looking for unique
solutions. This may not be so straightforward as 'nils'
don't exist at the lower level, which is where Sentient
eliminates duplicates.